requirements.txt
fastapi[standard]
uvicorn
gunicorn
# psycopg2-binary
oxyde
oxyde-admin
pydantic_settings
debugpy
httpx
redis
slowapi
tenacity
aiocache
python-jose[cryptography]
passlib